Refactoring code into repository (moving parts into internal)
